requirements-test.txt
# Testing dependencies for MiniTweet project
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.0.0
factory-boy>=3.2.0
coverage>=7.0.0
